build.sh.inc: Update the submodules to the specified version
Before this change, each submodule was being updated with the latest instead of the SHA set in this project.
This should apply to libiio and osc as well. We want to control which versions we're using for the dependencies of iio-osc and to osc itself.
Enhancement: update all modules in parallel

Signed-off-by: Dan Nechita <dan.nechita@analog.com>